Add source_path option to point pprof to source files
Currently pprof will look for source files only on the current directory
and its parents. This makes it hard to examine sources on jobs where
there are multiple source trees (eg from different libraries).

Add a variable to provide a search path for source files. It will default
to the cwd, so there will be no change in behavior by default.