Remove logging entries using a regexp
A previous commit attempted to handler mappings produces using the glog
package by matching the column on which the sentinel was found.

However, some tools generate the mapping using a single log entry,
where the prefix appears only on the first line.

Instead of matching by column, use a regexp to identify and match
prefixes introduced by the glog package.