Improve weblist assembly output (#124)
* Align file names in weblist disassembly

The disassembly displayed by weblist attempts to align file:line
information after each instruction, but two things interfere with this
alignment: 1) the instruction text often has tabs in it, which defeats
the length-based alignment and generally interacts poorly with HTML
rendering and 2) the instruction text often has special HTML
characters like < and > in it, but we pad the string after escaping
it, so the length is again not representative of how it will display.
For example, the following shows what this looks like for disassembly
that contains < and >:

                   .          .   41c634: cmp    $0x200,%rdx                               mgcsweepbuf.go:130
                   .          .   41c63b: jae    41c64f <runtime.(*gcSweepBuf).pop+0x4f> mgcsweepbuf.go:130
                   .          .   41c63d: mov    (%rax,%rdx,8),%rcx                        mgcsweepbuf.go:130
                   .          .   41c64f: callq  424800 <runtime.panicindex>             mgcsweepbuf.go:130
                   .          .   41c654: ud2                                              mgcsweepbuf.go:130

Fix these problems by replacing tab characters with spaces and padding
the string before escaping it. After this, the file names are properly
aligned:

                   .          .   41c634: cmp    $0x200,%rdx                               mgcsweepbuf.go:130
                   .          .   41c63b: jae    41c64f <runtime.(*gcSweepBuf).pop+0x4f>   mgcsweepbuf.go:130
                   .          .   41c63d: mov    (%rax,%rdx,8),%rcx                        mgcsweepbuf.go:130
                   .          .   41c64f: callq  424800 <runtime.panicindex>               mgcsweepbuf.go:130
                   .          .   41c654: ud2                                              mgcsweepbuf.go:130

* Separate discontiguous assembly blocks

Currently, weblist's disassembled output for each line simply lists
all instructions that are marked with that line number. This leads to
confusing disassembly because a lot of things look like straight-line
control flow when they aren't. For example, index panics look like
they always happen:

   .          .   41c62c: test   %al,(%rax)
   .          .   41c62e: and    $0x1ff,%edx
   .          .   41c634: cmp    $0x200,%rdx
   .          .   41c63b: jae    41c64f <runtime.(*gcSweepBuf).pop+0x4f>
   .          .   41c63d: mov    (%rax,%rdx,8),%rcx
   .          .   41c64f: callq  424800 <runtime.panicindex>
   .          .   41c654: ud2

In reality, 41c64f is at the end of the function, but it looks like we
call it immediately after a successful index operation.

Fix this by adding a vertical ellipses to separate blocks of assembly
that have other instructions between them. With this change, the above
looks like:

   .          .   41c62c: test   %al,(%rax)
   .          .   41c62e: and    $0x1ff,%edx
   .          .   41c634: cmp    $0x200,%rdx
   .          .   41c63b: jae    41c64f <runtime.(*gcSweepBuf).pop+0x4f>
   .          .   41c63d: mov    (%rax,%rdx,8),%rcx
                       ⋮
   .          .   41c64f: callq  424800 <runtime.panicindex>
   .          .   41c654: ud2