Fix unstable nodefraction handling. (#131)
There is an issue where decreasing nodefraction (which should produce
more nodes) ended up reducing the number of displayed nodes.  This is
because we have an upper limit on number of nodes to display and the
number of node-lets counts against that limit.  This can cause a problem
if a node with large number of node-lets shows up near the front of
the list of nodes (e.g., std::allocator_traits::allocate in a heapz
profile had 102 node-lets, which caused all subsequent nodes to be
dropped).

Fixed by limiting the count of charged node-lets per node to
maxNodelets (4), which is the maximum number of node-lets we
display per node.