Add a separate function to binutils to check if a file is an executable
Currently bu.Open() is used to detect if a file is an executable,
but it depends on the mapping, which is not available before opening
the input profiles.

Create a separate ExecutableBuildID function on the ObjTool interface that only
does a basic check to see if a binary is an executable, and extracts the build
id if available.

This produces better diagnostics if there are issues with the mappings, as we
will still recognize the file as a binary.