Allow passing in a custom HTTPServer (#341)
* Avoid opening browser when not in terminal

This also reverts e82ee9a, though only in spirit since the revert
did not apply cleanly.

This change may look cosmetic, but note that `plugin.UI` can be
passed in from the outside. This is important for embedding pprof
into an application for easy access to the web ui: with this change,
such an application can pass a non-Terminal `profile.UI` and avoids
the browser code.

* Allow passing in a custom HTTPServer

This allows embedding the pprof ui into an existing application.
See https://github.com/cockroachdb/cockroach/pull/24145 for an example.