Force saving the module name for unsymbolized frames
In some cases pprof isn't saving the module name in graph nodes, so
reports end up with nodes of the form '<unknown>'.

The module name is often stripped to allow symbolized frames from
different versions of a binary to be merge.

Force saving the objfile name if the name is unknown so we can print
some information for unsymbolized frames.