Look for symbols in debuginfo-compatible paths
Add paths compatible with gdb-style debuginfo when looking for
symbols. These are elfs containing just the symbols separated
from stripped binaries and are stored in paths like
.build-id/xx/xxxxxx.debug
where the directory is the first two characters of the build-id,
and the filename is the remaining.

On a typical Unix system these will be in
/usr/lib/debug, but also look in $PROF_BINARY_PATH as well.