Remove unneeded `Residual` from dotgraph test
I had unwittingly copied a test that was _specifically tesing_ Residual
edges (the name of the test set I copied was
`TestComposeWithTagsAndResidualEdge`).

In my test, I'm simply testing the printing of the _names_ of the edges,
and it's not relevant whether the edges are residual or not, so I've
removed it just to simplify the test.