Fix Windows test (after fixing Windows printing): properly escape `\` in paths
This PR adds proper escaping to dot strings, so that the backslash (`\`)
in windows paths will now print correctly in dot output. Previously, the
tests were incorreclty checking for an unescaped single `\` in the
output, which isn't a valid `dot` string, so this commit updates the dot
tests to expect the newly correct output.

Updates the path testing assertions in various test files