Add missing `escapeForDot()` to labels for function names (#564)
* Add missing `escapeForDot()` to labels for function names

In some programming languages, e.g. JuliaLang, function names can
contain arbitrary characters. These are represented via the string macro
`var"..."`, which allows constructing identifiers that wouldn't
otherwise parse.

These names are handled correctly by `pprof` in the FlameGraph view, but
before this commit, they would produce an invalid dot file.

This fixes the dot graph export for names that contain `"`.

* Add separate test for name escaping

* Apply `escapeStringForDot()` in more places, to cover more cases of potentially harmful string labels.

Remove mistaken `escapeStringForDot()` around tag names

* gofmt cleanups

* Cleanup: Remove commented out line

* Rename escapeForDot => escapeAllForDot; escapeStringForDot => escapeForDot

* Apply formatting suggestions from code review

* Remove unneeded `Residual` from dotgraph test

I had unwittingly copied a test that was _specifically tesing_ Residual
edges (the name of the test set I copied was
`TestComposeWithTagsAndResidualEdge`).

In my test, I'm simply testing the printing of the _names_ of the edges,
and it's not relevant whether the edges are residual or not, so I've
removed it just to simplify the test.

* Fix Windows test (after fixing Windows printing): properly escape `\` in paths

This PR adds proper escaping to dot strings, so that the backslash (`\`)
in windows paths will now print correctly in dot output. Previously, the
tests were incorreclty checking for an unescaped single `\` in the
output, which isn't a valid `dot` string, so this commit updates the dot
tests to expect the newly correct output.

Updates the path testing assertions in various test files

* Fix unexported comment in internal/graph/dotgraph.go

Co-authored-by: Alexey Alexandrov <aalexand@users.noreply.github.com>