Replace Base() with ObjAddr() in the plugin.ObjFile interface.
Modify the plugin.ObjFile interface by replacing the "Base() uint64" method
with an "ObjAddr(addr uint64) (uint64, error)" method. The latter computes
the objdump address corresponding to the given runtime address.

The change removes direct access to the base value, but enables us to delay
computing the base until it is actually needed.