Source listing handles missing binaries and addresses.
Some profiles can contain locations without an address, or reference
binaries/libraries that are not available. Treat these better when
producing source listings by using any source information present
directly in the profile.Location.

Details:

If a location does not have an address, create a synthetic address
that does not fall in a known mapping.

If a location references an object file that is not found, use any
file/line information baked directly in the location.

Tweaked generated html to clicking for source lines that have no
inlining/assembly to display.