Improve source listing when a binary is not available.
(1) If no address exists in a location, synthesize a fake address.

(2) If a location does not have a mapping, or the object file
identified by the mapping is not available, use the file/line/function
information embedded in the location istead.

This change avoids the problem of blank source listing when the
binary is not available, but when the profile already contains
embedded file/line info.