internal/report: error if filter argument has no matches
list, weblist, disasm, and peek all take a regexp argument that acts as
a filter.

Currently if this filter results in no matches we simply generate an
empty report, which can be confusing to users. Instead, explicitly treat
this as an error condition that is clearly reported to users.

Fixes #629