internal/graph: Escape tag labels correctly for dot
Fixes syntax errors when trying to render pprof profiles that have
double quotes in tags. These can be created with Go's pprof labels
feature, for example with:

    pprof.Labels("key", "label \"double quote\"\nline two")

Without this change, trying to render a diagram will fail with:

    Error: <stdin>: syntax error in line 5 near 'quote'

The problem is that the double quote ('"') was never escaped in the
label strings. This required me changing how multiple tags are
joined. Previously they were joined with the two character `\n`
sequence, which is a centered newline in dot. This changes it to
a single character newline "\n", which is then escaped as
left-justified strings "\l" in dot.

* Add a new graph test for labels.
* Add a test for joinLabels.
* Update the tests in driver which converts \n to \l.