Fix tagroot to properly format unitless numeric tags.
It's possible for the unit array to be empty for a unitless numeric tag.
In this case the tag value should be formatted properly but before this
change it was output empty.

This also coincidentally fixes #651. From the bug description it doesn't
look like exactly this issue but it does get fixed.

The added tests fail before the change and pass after.