internal/graph: keep Go type parameter ellipsis in dot
Go symbols with type parameters are reported in Go pprof profiles as
something like `main.Set[...]`.

Today, multilinePrintableName mishandles this by replacing each with a
newline, resulting in a very strange looking node.

Instead, replace ... with a unicode ellipsis so that it isn't replaced
with a newline just below.

Note that the full symbol name (which is reported by tools like perf) is
something like `[go.shape.string_0,go.shape.int_1]`. This case is still
not handled very well, but we likely want to strip this out anyways in
ShortenFunctionName or demangling as future work.

Fixes #705.