Add visual indication of inlined frames.
Mark frames sent to javascript with a boolean "Inlined" field. Use
this field to change the display by (1) adding an "(inlined)" marker
to the tooltip for the frame, and (2) by dropping any border between
the caller and the caller.

Document flame graph display, including coloring and visual indication
of inlining.