Do not overwrite main binary build ID when it's present. (#742)
When users specify the path to the main binary on the command line as an
easy way to make pprof find it, there is a possibility to make a mistake
and point to a wrong version of the binary. This will break
symbolization in unpredictable ways. I think we should only insert the
build ID if it is not present in the profile.

Co-authored-by: gmarin13 <gmarin13@users.noreply.github.com>