Remove the heuristic of recognizing build ID command line arg.
I am not entirely sure why this heuristic exists and we have -buildid
flag for this anyway. I suspect having this shortcut to specify the
build ID was for some reason more important in the past (more profiles
did not record the build ID of the main binary? not sure), but now it
seems very exotic to have to specify the build ID anyway.