Remove special coloring for self region in new flame graph view. (#782)
Previously, the self cost part of a flame graph box  was shown
with a tweaked color.

```
[<-- self --><-- rest of the cost -->]
             [child 1][child2][child3]
```

However this added clutter and was too subtle (I asked a couple of
colleagues and they hadn't noticed it).

This change removes this coloring. The self cost is still available in
the tooltip for each box and is also visible as the area to the left
of the first child.