Start using the new demangle.NoEnclosingParams demangling option.
github.com/ianlancetaylor/demangle now has new
`demangle.NoEnclosingParams` option that enables stripping the
parameters of a lambda's enclosing function, similar to how regular
functions' names are simplified with the `demangle.NoParams` option
pprof already uses.

E.g. we currently get
`foo_namespace::BarMethod(SomeArgType*)::$_0::operator()`, and with the
new option it will be `foo_namespace::BarMethod::$_0::operator()`.

It seems the right thing to start using that option: If there is a
function with a loop and then that loop is turned into an algorithm
function call with a lambda, it doesn't seem right that the name of that
lambda function suddenly gets parameters unstripped compared to the
straight loop version.